pyproject.toml
[tool.poetry]
name = "social-media"
version = "0.1.0"
description = ""
authors = ["Muhammad-Ramzan12 <mramzanali258@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = ">=5.1"
djangorestframework = ">=3.14"
psycopg2 = ">=2.9"
python-dotenv = ">=1.0"
drf-yasg = ">=1.21"
pytest = ">=7.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
